+++ src/codegen.hpp
@@ -24,7 +24,6 @@ extern vector<long long> calls_mod;
 // Register Allocation (Simple Tracker)
 // Maps Register ID (0-7) to a Symbol or NULL if free/trash.
 // Used to decide if we need to spill before Math Kernel calls.
-// 4=e, 5=f, 6=g are clobbered by Kernel.
 struct Symbol; // Forward declaration
 extern Symbol* reg_descriptors[8];
 
@@ -33,7 +32,7 @@ void emit(string opcode, long long arg, string comment);
 void emit(string opcode);
 void emit(string opcode, long long arg);
 void add_comment(string comment);
-void gen_const(int reg, long long value);
+void gen_const(int reg, BigInt value);
 void optimize_code();
 
 #endif
